pyproject.toml
[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[project]
name = "thatcat"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["quart", "msgpack", "jinja2", "cbor2", "ecdsa", "python-magic", "pytesseract"]
